executor updated for norm
