+++ Makefile
@@ -13,7 +13,7 @@ BUILTIN_DIR		=	builtin/
 BUILTIN_FILES	=	builtin cd echo env exit export pwd unset
 BUSINESS_DIR	=	business/
 BUSINESS_FILES	=	executor expander lexer main parser syntax_check_helper find_token_helper parser_helper \
-					add_type file_descriptors
+					add_type file_descriptors executor_helper
 ERROR_DIR		=	err/
 ERROR_FILES		=	err
 FREE_DIR		=	free/
